--- corecomponent/androidcomponent/src/main/res/navigation/navigation_sign_in.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/navigation_sign_in"
-    app:startDestination="@id/sign_in">
-
-    <fragment
-        android:id="@+id/sign_in"
-        android:name="com.techcafe.todone.auth.signin.SignInFragment"
-        android:label="SignIn" />
-
-</navigation>
