+++ corecomponent/androidcomponent/src/main/res/values-ja/strings.xml
@@ -2,6 +2,7 @@
     <string name="title_home">ホーム</string>
     <string name="title_now">ナウ</string>
     <string name="title_sign_in">サインイン</string>
+    <string name="title_Auth">認証</string>
     <string name="title_notifications">通知</string>
     <string name="item_projects">プロジェクト</string>
     <string name="item_settings">設定</string>
