+++ features/auth/src/main/java/com/techcafe/todone/auth/AuthActivity.kt
@@ -1,0 +1,17 @@
+package com.techcafe.todone.auth
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.techcafe.todone.auth.databinding.ActivityAuthBinding
+
+class AuthActivity : AppCompatActivity(R.layout.activity_auth) {
+
+    private lateinit var binding: ActivityAuthBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_auth)
+    }
+}
